refactor: move time provider
- move to external infrastructure
- add more options
