--- src/Server/Core/Application/Abstractions/Common/IDateTime.cs
@@ -1,8 +1,0 @@
-﻿namespace Application.Abstractions.Common;
-
-public interface ITimeProvider
-{
-    DateTime Now { get; }
-
-    DateTime UtcNow { get; }
-}
