+++ src/Server/Infrastructure/External/DependencyInjection.cs
@@ -1,0 +1,15 @@
+﻿using Application.Abstractions.Common;
+using External.TimeProvider;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace External;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddInfrastructure(this IServiceCollection services)
+    {
+        services.AddTransient<ITimeProvider, SystemTimeProvider>();
+
+        return services;
+    }
+}
