--- src/Server/Infrastructure/Persistence/Services/Common/SystemTimeProvider.cs
@@ -1,9 +1,0 @@
-﻿using Application.Abstractions.Common;
-
-namespace Persistence.Services.Common;
-
-public sealed class SystemTimeProvider : ITimeProvider
-{
-    public DateTime Now => DateTime.Now;
-    public DateTime UtcNow => DateTime.UtcNow;
-}
